NSOperationQueue
This is a slightly more complex method to running code off the main
thread, and it provides some additional control capabilities, like
running serially, concurrently, or with dependencies between tasks.
Operation queues are best suited to repetitive, well-defined
asynchronous tasks; for example, network calls or parsing.
Operation queues (NSOperationQueue) are available to manage a set of
tasks or operations (NSOperation).
An operation queue can specify how many operations can run
concurrently, can be suspended and restarted, and can cancel all
pending operations
The main difference here is that the NSOperationQueue is managing the
threads
This can be very important when tasks need to happen concurrently and
need to be managed to avoid overloading the system